pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "0.12.0"
poetry = "1.0.9"
requests = "2.24.0"
responses = "0.12.0"
gunicorn = "20.0.4"
pymongo = {extras = ["srv"], version = "^3.11.4"}

[tool.poetry.dev-dependencies]
pylint = "2.5.2"
pytest = "6.0.2"
selenium = "3.141.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"